# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "API for Postaway application",
    "title": "Postaway II"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "paths": {
...
    "/api/posts/": {
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Add new post with Image",
        "description": "User can upload new post with image. You can skip caption but can't skip Image(No use posting without Image... Duh!!!)",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string",
                    "required": false
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary",
                    "required": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Happens when image is not provied... You can't post without an Image. Try again with Image...!"
          }
        }
      },
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Filter posts by userId",
        "description": "User will get all the posts by specific user using userId as query parameter.",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric userId to get posts posted by a specific user"
          }
        ],
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found... No post found by specified user...!"
          }
        }
      }
    },
...
  }
}